feat(morpheme_http): Add support for HEAD requests
This commit adds a new method, `head()`, to the `MorphemeHttp` class in `morpheme_http.dart`. This method allows making HEAD requests to a specified URL. It accepts optional headers and cache strategy parameters.

Upgrade morpheme_http to version 3.2.0

This commit updates the version of `morpheme_http` in the `pubspec.yaml` file to 3.2.0. The library is a future-based package for making HTTP requests, modified from the http package.